+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Updates to beta services Data Guard and Prompt Guard.
+
 ## 5.2.0beta1 - 2024-10-23
 
 ### Added